Add comprehensive unit tests with 95% coverage and refactor module naming
Tests added:
- test_colors.py: 58 tests covering color generation, luminance, and contrast functions
- test_cli.py: 34 tests covering CLI parsing, validation, and integration
- conftest.py: Shared fixtures and test utilities
- Complete coverage of WCAG contrast compliance and color format parsing

Code improvements:
- Renamed pymp.py to colors.py for better semantic clarity
- Fixed numpy array casting bug in generate_readable_colors()
- Updated import statements and documentation references

Coverage achieved: 95% (176 statements, 8 missed)
- src/bgand256/colors.py: 98% coverage
- src/bgand256/cli.py: 94% coverage
- src/bgand256/__init__.py: 100% coverage

+++ src/bgand256/colors.py
@@ -22,7 +22,7 @@ def contrast_ratio(l1, l2):
 
 def generate_readable_colors(background_rgb):
     # Normalize background to [0,1] if needed
-    background_rgb = np.array(background_rgb)
+    background_rgb = np.array(background_rgb, dtype=float)
     if np.max(background_rgb) > 1.0:
         background_rgb /= 255.0
